refactor : index의 로직을 간략히. APP을 import 하도록 분리.
